Fix conversion problems when casting from void *
This fixes several problems that resulted from implicit conversion via the
item->v void* pointer.

In particular, float/int values caused headaches on 64 bit. 

Also removed a deallocation -- there is so much memory allocation here
that I fear it is safer to live with a memory leak than to deallocate
something which may not have been allocated in the first place :)