Enforce limit on minimum histogram noise value
A noise level which is a lot below the configured minimum noise level will
cause a very high SNR value. As a consequence the histogram screen is mostly
blue, and the green signal plot is barely visible.

This behaviour was observed with an Intel iw3945 wireless adapter which seems
not to report noise values -- it always reports a fixed value of
-127dBm. This is quite annoying since no matter which noise value is
configured, the screen remains mostly blue.

Therefore this patch enforces the configured minimum as a lower bound
when the noise value is below that value.

In addition it sets the configurable minimums down to the minimum
possible values (signed 8 bit arithmetic of the wireless tools).