Fix bug in displaying the operation mode
The array for the opmodes was out-of-date, since access to it was
unchecked, a modern card in mode 'monitor' would have caused a
segfault.
Fixed by consolidating the code to display the operation mode in
a separate function. In addition, the dump_parameters() function
to display interface parameters has been moved to iw_if.c