Improved saving of round-robin data
The current implementation creates a FIFO out of an array by shifting 1023
entries on each insertion. Given that sizeof(struct iw_stat) = 24, this means
copying approximately 24 kB in memory for an insertion of just 24 bytes.

This patch therefore adds a simple circular buffer to avoid the frequent
expensive copies. It is again based on a fixed-size array; the circular
semantics are achieved by using an unsigned counter and a simple hash function
to take care of the array wrap-around. Set/get functions are provided.