Sorted list of access points
When operating as an access point, the presence of other access points
indicates a source of interference. As an extreme case, in a densely
populated area, 10 other access points were counted in an apartment
block - which is more than 3 times too much for the ISM band.

Sorting the access point list in descending order of signal strength
now presents the biggest sources of interference first.