Enforce minimum screen size
To display the screen clearly, minimally 25 lines by 80 columns are
needed, otherwise parts of the screen are missing and/or distorted.

This patch enforces the minimum size by issuing a warning about the
required screen size. In addition, magic numbers referring to screen
sizes have been replaced with symbolic constants.