Ignore unknown configuration items
When parsing the rc file, it is not strictly necessary to throw an error
when encountering an unknown identifier. In particular, unknown identifiers
can also be caused by outdated configuration files, such as:

 wavemon: parse error in ~/.wavemonrc, line 4: unknown identifier 'linear_scale'

This patch causes unknown identifiers to be ignored - a warning message is shown
instead. Thanks to Nelson A. de Oliveira for raising this.