Bug-fix: initialisation of histogram data
The bug fixed by this patch affects the initialisation of the 'range'
element of the struct iw_stat. This structure element contains interface
data and therefore need not be updated on each cycle.

The common code to update histogram/info screen data structures is shared
between the info screen and histogram screen source files.

However, the initialisation was wrongly done in scr_info(): hence the iw_stat
data structure was not properly initialised when using the histogram screen.

Fixed by moving the initialisation of the 'range' element more appropriately
into the shared function 'reinit_on_changes()'.