Import helper functions to support scan events
Parsing scan results is not simple, supporting it requires to import a
substantial part of iwlib. In order to update the access point list window
to phase out the defunct SIOCGIWAPLIST ioctl, a large part of iwlib from
wireless tools 30 is imported, enough to support scan events under different
circumstances. Many thanks to Jean Tourrilhes for providing these functions,
this has helped a great deal to avoid reinventing the wheel.