Scan window: perform scan operation in child process
Gathering the scan data can take multiple seconds, during this time
keyboard input will not be read. To increase responsiveness, fork a
child process for the scan operation, which is killed when changing
the window. To avoid zombie processes, ignore SIGCHLD; also ignore
TSTP since data gathering goes on in the background.
Using a delay of 0.2s between updates was found to be sufficient.