Restrict continual sampling to info/histogram screens
To avoid interference with the scan operation, restrict periodic sampling of
wireless statistics to the info/histogram screens. Polling is started upon
opening the screen, and stopped when leaving it. A byproduct is that state
is automatically up-to-date when e.g. configuration items are updated.
Also some minor code clean-up.