Support common options even if wireless extensions are absent
If a system has no wireless interfaces, or if wireless extensions for its
wireless interfaces are absent (no /proc/net/wireless), wavemon previously
aborted with an error message, even when given the -v/-h options.

Changed the logic so that the error message appears last and is more
informative to a user. Also allow to run -v/-h or -v/-d together.