Keyboard: work-around for PF1..4 escape sequences
Fix handling of function keys on terminals that send the PF1..4 escape
sequences instead of the advertised terminfo keycodes for F1..4. This problem
was observed on aterm, which uses 'rxvt' terminfo entry, where F1..F4 are
listed as \E1{1..4}~, but the generated key sequence is actually \EOP..\EOR
(vt100 PF1..4). Since ncurses can not match the keycode with the terminfo
entry, fix this particular problem by re-interpreting the escape sequence
as F1..4. Tested to work on aterm and to not interfere on other terminals.