fix bug displaying junk in the interface list pref dialog
conf.c, line 317 assigns the current address of the if_list variable
to another variable.

        item->list      = if_list;

since the list is later freed,realloced,etc,its address may and does
change, so the stored pointer in item->list points to junk.

issue experienced after starting wavemon linked against musl libc
for the second time (supposedly after creating a config file on
the first run).
the issue may not be experienced with other malloc implementations
(for example glibcs) with "luck", but it's still undefined behaviour
to access a pointer to freed storage.

there are 2 possibly fixes; one is to store a pointer to the list
rather than the list itself in the conf list, however this involves
making ugly changes and dealing with char*** in a lot of places.
the solution i went for is to make if_list a bounded array whose
address never changes (only those of its items).
the list is set to a maximum of 64 entries, so this wastes at most
64*(sizeof (char*)) bytes, which is 512 bytes on 64bit cpus.
having more interfaces would not be practical, as you can only iterate
through the interface list in a forwards/backwards fashion rather than
direct access.

additionally a call to strncpy was replaced with snprintf, as strncpy
may not zero-terminate given a long enough input string, and otherwise
does unnecessary padding. (see man 3p strncpy for details)