Scan screen: add error handling
This fixes a bug of omission, not handling the nl80211 error code.  It is
fixed so far for the scan screen, the other commands still ignore errors. This
will have to be revisited, to see if that causes problems, or if it is ok
to continue just with empty results.

The scan code still has a bug: when doing too many wild moves of screen
resizing, it will hang on 'Waiting for scan data'. Deferring this until later,
since running out of time.  Otherwise the scan code now works more reliably
and predictably.