Info Screen: tentatively move screen-update code
Running the display updates from within the pthread is problematic.
This may be due that the shared screen/terminal is now manipulated
from two different threads. It leads to interference and ANSI escape
sequences being printed on the screen.
This patch moves the display updates back into the main thread.
Some more testing is required to see if this has any side effects.