Scan screen: restore interface state on exit
Scanning can not be done when the interface is in a 'down' state.
Bringing the interface up within wavemon may interfere with other
programs that rely on the interface state to be down initially (#46).

Hence restore 'down' state on exit if the interface has to be brought
up in order to enable scanning.