Info Screen: fix and revise use of locks
This fixes several problems with the current use of the mutex,
which was poorly implemented, did not protect all the data.
The biggest problem was that the lock was taken too long (during
two UI-updating functions), which caused the screen to lock up
when the terminal was resized.
The update now just swaps pointers within the critical section.