Scan/Info Screen: use error-checking mutexes
This fixes a long-standing problem: on a SIGWINCH, the loop function
is abandoned, and fini() is called. If the loop function held on to
a mutex, this causes a deadlock condition. Normal mutexes can not be
unlocked if the current thread is not the owner.
The solution is to use error-checking mutexes (and ignoring the error
if the mutex is owned by a different thread). Since both screens cancel
via pthread_cancel, I think that a robust mutex is not required here.
Resolves #96.