History Screen: remove support for noise levels
The changes in #97 removed support for noise levels in the history
plot, this commit removes the remaining code. The rationale behind
this is that few cards even report noise levels; those that I have
seen only report a constant value. Hence the same information can
also be represented by a simpler and single signal graph.