ncurses: handle white backgrounds, too (#119)
White backgrounds broke with the support of transparency because only
the default bg color was considered. The fg color was still hardocded
to white, resulting in white on white e.g. for the config screen. So,
the default fg color needs also to be considered. Luckily, color pair 0
is set to the default colors after use_default_colors(), so we simply
need to map the index for the standard color accordingly. This will also
work when transparency is disabled because ncurses initializes color
pair 0 to white on black when not calling use_default_colors() which is
exactly what we want in this case.

Co-authored-by: Wolfram Sang <wsa@kernel.org>