iw_nl80211: don't block when receiving msgs
Info screen waits for the first data to arrive. This stalls sometimes on
my machine because the netlink command does not complete for some
reason. So, set the socket to non-blocking and try again next cycle if
no data is available. This needs a small version bump for libnl from
3.2 to 3.2.22 because only since then the call to nl_recvmsgs() returns
-NLE_AGAIN. Fixes #120.